requirements.txt
filelock==3.19.1
fsspec==2025.9.0
Jinja2==3.1.6
MarkupSafe==2.1.5
mpmath==1.3.0
mss==10.1.0
networkx==3.5
numpy==2.3.4
opencv-python==4.12.0.88
pillow==11.3.0
PyAutoGUI==0.9.54
pynput==1.8.1
PyYAML==6.0.3
setuptools==70.2.0
sympy==1.14.0
torch==2.9.0+cpu
torchvision==0.24.0+cpu
typing_extensions==4.15.0
ultralytics==8.3.225
